requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0

djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.2.0
django-guardian=2.4.0

celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

django-filter>=23.2
python-dotenv>=1.0.0
Pillow>=10.0.0
markdown>=3.4.4
django-extensions>=3.2.3

python-decouple
mysqlclient

drf-spectacular